Split helpers out of main FeaturedCard file
