+++ content/webapp/components/FeaturedCard/index.tsx
@@ -1,7 +1,5 @@
 import { FunctionComponent, PropsWithChildren } from 'react';
 
-import { ImageType } from '@weco/common/model/image';
-import { Label } from '@weco/common/model/labels';
 import linkResolver from '@weco/common/services/prismic/link-resolver';
 import { transformImage } from '@weco/common/services/prismic/transformers/images';
 import { font, grid } from '@weco/common/utils/classnames';
@@ -12,18 +10,13 @@ import { PaletteColor } from '@weco/common/views/themes/config';
 import DateRange from '@weco/content/components/DateRange/DateRange';
 import StatusIndicator from '@weco/content/components/StatusIndicator/StatusIndicator';
 import { Article } from '@weco/content/services/wellcome/content/types/api';
-import { ArticleBasic } from '@weco/content/types/articles';
-import { BookBasic } from '@weco/content/types/books';
-import { Card } from '@weco/content/types/card';
-import { EventSeries } from '@weco/content/types/event-series';
-import { EventBasic } from '@weco/content/types/events';
 import { ExhibitionBasic } from '@weco/content/types/exhibitions';
-import { Guide } from '@weco/content/types/guides';
-import { Link } from '@weco/content/types/link';
-import { Page } from '@weco/content/types/pages';
-import { Season } from '@weco/content/types/seasons';
-import { SeriesBasic } from '@weco/content/types/series';
 
+import {
+  convertCardToFeaturedCardProps,
+  convertItemToFeaturedCardProps,
+  PartialFeaturedCard,
+} from './FeaturedCard.helpers';
 import {
   DateWrapper,
   FeaturedCardCopy,
@@ -33,12 +26,6 @@ import {
   FeaturedCardShim,
   FeaturedCardWrap,
 } from './FeaturedCard.styles';
-
-type PartialFeaturedCard = {
-  image?: ImageType;
-  labels: Label[];
-  link: Link;
-};
 
 type Props = PartialFeaturedCard & {
   background: PaletteColor;
@@ -94,54 +81,6 @@ const FeaturedCard: FunctionComponent<PropsWithChildren<Props>> = ({
   );
 };
 
-export function convertCardToFeaturedCardProps(
-  item: Card
-): PartialFeaturedCard {
-  return {
-    // We intentionally omit the alt text on promos, so screen reader
-    // users don't have to listen to the alt text before hearing the
-    // title of the item in the list.
-    //
-    // See https://github.com/wellcomecollection/wellcomecollection.org/issues/6007
-    image: item.image && {
-      ...item.image,
-      alt: '',
-    },
-    labels: item.format ? [{ text: item.format.title }] : [],
-    link: { url: item.link || '', text: item.title || '' },
-  };
-}
-
-export function convertItemToFeaturedCardProps(
-  item:
-    | ArticleBasic
-    | ExhibitionBasic
-    | Season
-    | Page
-    | EventSeries
-    | BookBasic
-    | SeriesBasic
-    | EventBasic
-    | Guide
-): PartialFeaturedCard {
-  return {
-    image: item.promo?.image && {
-      ...item.promo.image,
-      // We intentionally omit the alt text on promos, so screen reader
-      // users don't have to listen to the alt text before hearing the
-      // title of the item in the list.
-      //
-      // See https://github.com/wellcomecollection/wellcomecollection.org/issues/6007
-      alt: '',
-    },
-    labels: item.labels,
-    link: {
-      url: linkResolver(item),
-      text: item.title,
-    },
-  };
-}
-
 type FeaturedCardArticleProps = {
   article: Article;
   background: PaletteColor;
@@ -191,6 +130,7 @@ export const FeaturedCardArticle: FunctionComponent<
     text: article.title,
   };
   const labels = [{ text: article.format.label }];
+
   return (
     <FeaturedCard
       image={image}
@@ -229,3 +169,4 @@ export const FeaturedCardExhibition: FunctionComponent<
 };
 
 export default FeaturedCard;
+export { convertCardToFeaturedCardProps, convertItemToFeaturedCardProps };
